Also trigger on fas.role.update.
So, I applied to the gitsupybot-fedora group, and just sat there.
Later, @ianweller sponsored me which really added me to the group.  It
triggered a ``fas.role.update`` message from fas.

However, hosted03 never got a new run of fas client and thus never
learned that I was now in this group (and should have privs to make
releases of stuff).

This should fix that.